requirements.txt
deepface
pg8000
python-dotenv
pillow
tf-keras